tune style_modifier per style from eval ratings, allow fractional values
Applied recommended style_modifier values (Ukiyo-e, Expressionism, Vector
Art, Doodle Art, Ink, Oil Painting, Chinese art, Ghibli, Cyberpunk,
Illustration) derived from balanced ratings in the qr_eval_brightness_styles
dashboard. Half-step values only take effect if style_modifier survives as
a float end-to-end, so also removed the frontend Math.round() and widened
the backend query param/type hints from int to float.

+++ api/controllers/generate_controller.py
@@ -93,7 +93,7 @@ async def predict(
     style_title: str,
     style_loras: str = "[]",
     qr_weight: int = 0,
-    style_modifier: int = 0,
+    style_modifier: float = 0,
 ):
     try:
         # --------------------------------- CHECK FUNDS ------------------------------- #
