+++ api/main.py
@@ -112,7 +112,7 @@ async def generate_endpoint(
     style_loras: Annotated[str, Query(max_length=2000)] = "[]",
     seed: Annotated[int, Query(ge=-1)] = -1,
     qr_weight: Annotated[int, Query(ge=-2, le=2)] = 0,
-    style_modifier: Annotated[int, Query(ge=-2, le=2)] = 0,
+    style_modifier: Annotated[float, Query(ge=-2, le=2)] = 0,
     current_user: dict = Depends(get_current_user),
 ):
     return await predict(
